requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
chromadb
pydantic
python-dotenv
pillow